Organize items_card
