+++ lib/src/features/inventory/widgets/inventory_details/inventory_details.dart
@@ -2,7 +2,7 @@ import 'package:file_picker/_internal/file_picker_web.dart';
 import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:librarian_app/src/features/inventory/widgets/inventory_details/items_card/create_items_dialog.dart';
+import 'package:librarian_app/src/features/inventory/widgets/inventory_details/items_card/create_items/create_items_dialog.dart';
 import 'package:librarian_app/src/widgets/fields/checkbox_field.dart';
 import 'package:librarian_app/src/widgets/input_decoration.dart';
 import 'package:librarian_app/src/features/inventory/models/updated_image_model.dart';
@@ -12,7 +12,7 @@ import 'package:librarian_app/src/features/inventory/providers/selected_thing_pr
 import 'package:librarian_app/src/features/inventory/providers/thing_details_provider.dart';
 import 'package:librarian_app/src/features/inventory/providers/things_repository_provider.dart';
 import 'package:librarian_app/src/features/inventory/widgets/inventory_details/categories_card.dart';
-import 'package:librarian_app/src/features/inventory/widgets/inventory_details/items_card/item_details_dialog.dart';
+import 'package:librarian_app/src/features/inventory/widgets/inventory_details/items_card/item_details/item_details_dialog.dart';
 import 'package:librarian_app/src/features/inventory/widgets/inventory_details/items_card/items_card.dart';
 import 'package:librarian_app/src/features/inventory/widgets/inventory_details/thing_image_card/thing_image_card.dart';
 import 'package:librarian_app/src/utils/media_query.dart';
